Description
The expression x^y is negative if x and y have opposite signs. We know that the leftmost bit is 0, for a positive number and 1 for a negative number. The XOR(pronounced as X R) operator will set the leftmost bit of output is zero for similar sign integers, and for opposite integers, it will set the leftmost bit as 1.
Example:
Let x=4     0000 00100
Let y=5     0000 00101
x ^ y         0000 00001  It is a positive number.

Now let x=4, and y=-5
0000 00100   =  4
1111 11011  =  -5 (2's complement form)
---------------
1111 11111  =  x^y  It is a negative number